fix: fixed bug causing movies not to be saved in database in API 23 and above. Th issue was related to the GSON library as discussed under - https://github.com/google/gson/issues/648

+++ app/src/main/java/com/rowland/moviesquire/ui/activities/MainActivity.java
@@ -67,7 +67,7 @@ public class MainActivity extends BaseToolBarActivity implements MainFragment.IM
             if (savedInstanceState != null) {
                 // Recover assets
                 mSelectedPosition = savedInstanceState.getInt(SELECTED_POSITION_KEY);
-                mMovieId = savedInstanceState.getInt(SELECTED_MOVIE_KEY);
+                mMovieId = savedInstanceState.getLong(SELECTED_MOVIE_KEY);
                 // Set positions as selected
                 onMovieSelected(mMovieId, mSelectedPosition);
                 return;
